v2 Stage A move #1: extract read-only state utilities into state/reader.mjs
Creates the hooks/lib/{state,policy,observability}/ subdirectory shape from
the v2 redesign, then performs one safe extraction: 12 read-only symbols
move from mpl-state.mjs into state/reader.mjs. mpl-state.mjs becomes a
facade — every previously-exported symbol is re-exported at the same name,
so all 33+ caller import sites keep working without edit.

Moved (read-only, no behavior change):
  readState · isMplActive · detectStateDrift · checkConvergence
  migrateLegacyExecutionState · applyMigrationChain (internal)
  readPersistedSchemaVersion (internal, used by writer's H8 guard)
  STATE_DIR · STATE_FILE · CURRENT_SCHEMA_VERSION
  MAX_AMBIGUITY_HISTORY · LEGACY_EXECUTION_STATE_PATH

Untouched (deliberate — future v2 moves):
  writeState · initState · cleanPipelineScope · archivePreviousRun
  recordFixLoopHistory · deriveLegacyGateBooleans · normalizeBlockedHookState
  mcp-server/src/lib/state-manager.ts (v2 vs v7 schema skew)

Verify: 1,629 tests pass · 0 regressions · facade re-export covers every
existing public symbol.
